# src/Endpoint/VMware.ScriptRuntimeService.APIGateway/swagger.yaml
openapi: 3.0.1
info:
  title: Script Runtime Service for vSphere
  description: "# Script Runtime Service API\r\n\r\nScript Runtime Service for vSphere (SRS) allows running PowerShell and PowerCLI scripts. SRS is a VC add-on that is deployed separately from VCSA. SRS can be accessed via REST API that allows you to create PowerShell instances and run PowerShell and PowerCLI scripts within. No Connect-VIServer is required to run PowerCLI against VC(s) SRS is registered to.\r\n\r\n## Authetication\r\n\r\nSRS uses VC SSO as Identity and Authentication Server. Two types of authentication are supported. SIGN and Basic. SIGN authentication is purposed for Service-To-Service access to SRS resources. For the convenience of the end-users, SRS supports basic authentication passing username and password which are used to acquire SAML HoK token for SRS solution. When basic is used SRS exchanges the username and password for SAML HoK token from the SSO server. SRS uses the SAML token to Connect PowerCLI to VC services in further operations. \r\nOn successful authentication SRS returns API Key which is required to authorize further SRS API calls."
  version: '1.0'
  x-vmw-vapi-codegenconfig:
    package-name: com.vmware.srs
paths:
...
  /api/auth/login:
    post:
      tags:
        - authentication
      summary: Exchanges client credentials or SIGN token for SRS access key
      description: "Uses vCenter Server SSO as Identity and Authentication Server.\r\nTwo types of authentication are supported SIGN and Basic.\r\nWhen Basic authentication is used, the service exchanges username and password for SAML Token from the vCenter Server SSO.\r\nWhen SIGN authentication is used, the service exchanges the SSO SAML token from the SIGN token for another SAML token on behalf of the principal from the vCenter Server SSO.\r\nOn successful authentication with the SSO, the service returns **X-SRS-API-KEY** token in the response headers. **X-SRS-API-KEY** token authorizes access to the service resources. \r\nThe service associates **X-SRS-API-KEY** token to the acquired SSO SAML token. The associated SSO SAML token authenticates PowerCLI to the vCenter services."
      operationId: login
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - basicAuth: []
        - signAuth: []
      x-vmw-vapi-servicename: authentication
      x-vmw-vapi-methodname: login
      x-vmw-doc-operation: login
    x-vmw-doc-resource: session
    x-vmw-doc-domain: authentication
  /api/auth/logout:
    post:
      tags:
        - authentication
      summary: Revokes SRS access key
      description: "The service revokes **X-SRS-API-KEY** and deletes all associated non-active **runspace** resources.\r\nActive runspaces are deleted immediately after the completion of the scripts they run."
      operationId: logout
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: authentication
      x-vmw-vapi-methodname: logout
      x-vmw-doc-operation: logout
    x-vmw-doc-resource: session
    x-vmw-doc-domain: authentication
  /api/runspaces:
    post:
      tags:
        - runspaces
      summary: Starts a runspace creation
      description: "### Create a runspace\r\nRunspace creation and preparation time depends on the requested runspace details.\r\nIf a connection to the vCenter Servers is requested, the operation creates a PowerShell instance, loads PowerCLI modules, and connects PowerCLI to the vCenter Servers.\r\n### Returns\r\nWhen request is accepted **202 Accepted** - response code, with **Location** header is returned in the response that leads you to the **runspace** resource. The **runspace** resource is in **creation** state initially."
      operationId: create.runspace
      requestBody:
        description: Desired runspace resource.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Runspace'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: create
      x-vmw-doc-operation: create.runspace
    get:
      tags:
        - runspaces
      summary: List all runspaces
      description: "### List all runspaces\r\n### Returns\r\nReturns a list of your runspaces."
      operationId: list.runspaces
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: list
      x-vmw-doc-operation: list.runspaces
    x-vmw-doc-resource: Runspace
    x-vmw-doc-domain: runspaces
  '/api/runspaces/{id}':
    get:
      tags:
        - runspaces
      summary: Retrieve a runspace
      description: "### Retrieve a runspace\r\nRetrieves the details of a runspace. One only needs to supply the unique runspace identifier returned on the runspace creation to retrieve runspace details.\r\n### Returns\r\nReturns a **runspace** resource instance if a valid identifier was provided.\r\nWhen requesting the Id of a runspace that has been deleted or doesn't exist **404 NotFound** is returned."
      operationId: get.runspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: get
      x-vmw-doc-operation: get.runspace
    delete:
      tags:
        - runspaces
      summary: Deletes a runspace
      description: "### Deletes a runspace\r\nDeletes the PowerShell instance that is prepresented by this **runspace** resource.\r\nRunning script in the PowerShell instance won't prevent the operation.\r\n### Returns\r\nWhen requesting the Id of a runspace that has been deleted or doesn't exist **404 NotFound** is returned."
      operationId: delete.runspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: delete
      x-vmw-doc-operation: delete.runspace
    x-vmw-doc-resource: Runspace
    x-vmw-doc-domain: runspaces
...
components:
  schemas:
...
    ErrorDetails:
      type: object
      properties:
        code:
          type: integer
          description: Code of the internal service error. It is useful when error is reported to service maintainers.
          format: int32
        error_message:
          type: string
          description: Error message.
          nullable: true
        details:
          type: string
          description: Details of the error.
          nullable: true
      additionalProperties: false
      description: ErrorDetails object gives information for an API error.
      readOnly: true
    RunspaceState:
      enum:
        - ready
        - active
        - creating
        - error
      type: string
      description: RunspaceState defines possible runspace states.
      readOnly: true
    Runspace:
      type: object
      properties:
        name:
          type: string
          description: Name of the runspace. It is optional to give a name of the runspace on create request. If name was not specified on runspace creation the field has null value.
          nullable: true
        id:
          type: string
          description: Unique identifier for the object.
          nullable: true
          readOnly: true
        state:
          $ref: '#/components/schemas/RunspaceState'
        error_details:
          $ref: '#/components/schemas/ErrorDetails'
        run_vc_connection_script:
          type: boolean
          description: "Boolean that indicates whether PowerCLI connection has to be initialized when a Runspace creation is requested.\r\n\r\nIt is optional to request VCenter Servers connection to be initialized on create request.\r\nIf true is specified PowerCLI Connect-VIServer cmdlet will be called in the create PowerShell instance.\r\nConnect-VIServer uses the authorized user's SSO SAML token to connect to the vCenter Servers.\r\nConnection to all linked vCetner Servers will be established. When requested the runspace creation will be slower\r\nwith the time needed for PowerCLI modules loading and Connect-VIServer execution time. _vc_connection_script_state\r\nfield will hold state of the script completion. In case errors have occured during connect script execution\r\n_vc_connection_script_error_records will be populated.\r\n\r\nDefault value is false."
        vc_connection_script_id:
          type: string
          description: "Id of the vc connection script if it was requested on the runspace create request.\r\nYou can use this id to retrieve all the details of the script execution.\r\nThe script could fail and the details about the script failure would be available in the scriptexecutions API."
          nullable: true
          readOnly: true
        creation_time:
          type: string
          description: The time at which the object was created. String representing time in format ISO 8601.
          format: date-time
          readOnly: true
      additionalProperties: false
      description: "Runspace object allows you to create and delete isolated PowerShell instance for your script executions.\r\nThe API allows you to create, read, and delete runspaces."
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      name: X-SRS-API-KEY
      in: header
    basicAuth:
      type: http
      scheme: basic
    signAuth:
      type: http
      scheme: SIGN
security:
  - apiKeyAuth: []
  - basicAuth: []
  - signAuth: []